+++ Plutarch/Prelude.hs
@@ -273,6 +273,12 @@ module Plutarch.Prelude (
   punrollBound',
   punrollUnbound,
   punrollUnboundWhole,
+
+  -- * Validation
+  PValidateData (..),
+  Don'tValidate (..),
+  DeriveNewtypePValidateData (..),
+  pparseData,
 ) where
 
 import Plutarch.Builtin.Array
@@ -297,6 +303,7 @@ import Plutarch.Internal.Numeric
 import Plutarch.Internal.Ord
 import Plutarch.Internal.Other
 import Plutarch.Internal.PLam
+import Plutarch.Internal.Parse
 import Plutarch.Internal.PlutusType
 import Plutarch.Internal.Quantification
 import Plutarch.Internal.ScottEncoding