+++ plutarch-ledger-api/src/Plutarch/LedgerApi/Utils.hs
@@ -43,7 +43,6 @@ module Plutarch.LedgerApi.Utils (
 import Data.Kind (Type)
 import GHC.Generics (Generic)
 import Generics.SOP qualified as SOP
-import Plutarch.Internal.Parse (PValidateData)
 import Plutarch.Internal.PlutusType (PlutusType (pcon', pmatch'))
 import Plutarch.Prelude
 import Plutarch.Unsafe (punsafeCoerce)
@@ -151,6 +150,8 @@ data PRationalData s = PRationalData
   deriving
     ( -- | @since 3.3.0
       PlutusType
+    , -- | @since wip
+      PValidateData
     )
     via (DeriveAsDataStruct PRationalData)