+++ plutarch-ledger-api/src/Plutarch/LedgerApi/V1/Contexts.hs
@@ -36,6 +36,8 @@ data PScriptPurpose (s :: S)
   deriving
     ( -- | @since 3.3.0
       PlutusType
+    , -- | @since wip
+      PValidateData
     )
     via (DeriveAsDataStruct PScriptPurpose)