+++ plutarch-ledger-api/src/Plutarch/LedgerApi/V1/Tx.hs
@@ -4,6 +4,7 @@
 module Plutarch.LedgerApi.V1.Tx (
   PTxId (..),
   PTxOutRef (..),
+  txIdByteSize,
 ) where
 
 import GHC.Generics (Generic)
@@ -48,6 +49,29 @@ deriving via
 -- | @since 3.4.0
 instance PTryFrom PData (PAsData PTxId)
 
+{- | Checks that we have a 'PTxId' of valid length. The underlying
+'PByteString' must be exactly 32 bytes, as Cardano transactions are hashed
+with BLAKE2b-256.
+
+@since wip
+-}
+instance PValidateData PTxId where
+  pwithValidated opq x =
+    -- FIXME: Why is it wrapped in Constr 0?
+    pmatch (pasConstr # opq) $ \(PBuiltinPair constrIdx fields) ->
+      pif
+        ((constrIdx #== 0) #&& ((plength # fields) #== 1))
+        ( plet (plengthBS #$ pfromData $ pparseData @PByteString $ phead # fields) $ \bsSize ->
+            pif
+              (bsSize #== txIdByteSize)
+              x
+              perror
+        )
+        perror
+
+txIdByteSize :: forall (s :: S). Term s PInteger
+txIdByteSize = 32
+
 {- | Reference to a transaction output, with an index referencing which exact
 output we mean.
 
@@ -88,3 +112,23 @@ deriving via
 
 -- | @since 3.4.0
 instance PTryFrom PData (PAsData PTxOutRef)
+
+{- | Checks that we have a valid 'PTxOutRef'. The underlying 'PTxId' must be
+exactly 32 bytes, as Cardano transactions are hashed with BLAKE2b-256, and
+the output index must be a non-negative 'PInteger'.
+
+@since wip
+-}
+instance PValidateData PTxOutRef where
+  pwithValidated opq x =
+    pmatch (pasConstr # opq) $ \(PBuiltinPair constrIdx fields) ->
+      pif
+        ((constrIdx #== 0) #&& ((plength # fields) #== 2))
+        ( pwithValidated @PTxId (ptryIndex 0 fields) $
+            plet (pasInt # ptryIndex 1 fields) $ \outIdx ->
+              pif
+                (outIdx #< 0)
+                perror
+                x
+        )
+        perror