+++ plutarch-ledger-api/src/Plutarch/LedgerApi/Value/CurrencySymbol.hs
@@ -51,6 +51,26 @@ instance PLiftable PCurrencySymbol where
 -- | @since 3.4.0
 instance PTryFrom PData (PAsData PCurrencySymbol)
 
+{- | Checks that we have a 'PCurrencySymbol' of valid length. The underlying
+'PByteString' must be exactly 28 bytes, as Cardano minting policies are hashed
+with BLAKE2b-224.
+
+@since wip
+-}
+instance PValidateData PCurrencySymbol where
+  pwithValidated opq x =
+    plet (plengthBS #$ pasByteStr # opq) $ \bsSize ->
+      pif
+        ((bsSize #== pnonAdaCurrencySymbolByteSize) #|| (bsSize #== padaCurrencySymbolByteSize))
+        x
+        perror
+
+pnonAdaCurrencySymbolByteSize :: forall (s :: S). Term s PInteger
+pnonAdaCurrencySymbolByteSize = 28
+
+padaCurrencySymbolByteSize :: forall (s :: S). Term s PInteger
+padaCurrencySymbolByteSize = 0
+
 {- | The 'PCurrencySymbol' of the Ada currency.
 
 @since 2.1.1